+++ lib/l10n/clip_localizations.dart
@@ -1,0 +1,156 @@
+
+import 'dart:async';
+
+// ignore: unused_import
+import 'package:flutter/foundation.dart';
+import 'package:flutter/widgets.dart';
+import 'package:flutter_localizations/flutter_localizations.dart';
+import 'package:intl/intl.dart' as intl;
+
+import 'clip_localizations_ar.dart';
+import 'clip_localizations_en.dart';
+
+/// Callers can lookup localized strings with an instance of ClipLocalizations returned
+/// by `ClipLocalizations.of(context)`.
+///
+/// Applications need to include `ClipLocalizations.delegate()` in their app's
+/// localizationDelegates list, and the locales they support in the app's
+/// supportedLocales list. For example:
+///
+/// ```
+/// import 'l10n/clip_localizations.dart';
+///
+/// return MaterialApp(
+///   localizationsDelegates: ClipLocalizations.localizationsDelegates,
+///   supportedLocales: ClipLocalizations.supportedLocales,
+///   home: MyApplicationHome(),
+/// );
+/// ```
+///
+/// ## Update pubspec.yaml
+///
+/// Please make sure to update your pubspec.yaml to include the following
+/// packages:
+///
+/// ```
+/// dependencies:
+///   # Internationalization support.
+///   flutter_localizations:
+///     sdk: flutter
+///   intl: any # Use the pinned version from flutter_localizations
+///
+///   # rest of dependencies
+/// ```
+///
+/// ## iOS Applications
+///
+/// iOS applications define key application metadata, including supported
+/// locales, in an Info.plist file that is built into the application bundle.
+/// To configure the locales supported by your app, you’ll need to edit this
+/// file.
+///
+/// First, open your project’s ios/Runner.xcworkspace Xcode workspace file.
+/// Then, in the Project Navigator, open the Info.plist file under the Runner
+/// project’s Runner folder.
+///
+/// Next, select the Information Property List item, select Add Item from the
+/// Editor menu, then select Localizations from the pop-up menu.
+///
+/// Select and expand the newly-created Localizations item then, for each
+/// locale your application supports, add a new item and select the locale
+/// you wish to add from the pop-up menu in the Value field. This list should
+/// be consistent with the languages listed in the ClipLocalizations.supportedLocales
+/// property.
+abstract class ClipLocalizations {
+  ClipLocalizations(String locale) : localeName = intl.Intl.canonicalizedLocale(locale.toString());
+
+  // ignore: unused_field
+  final String localeName;
+
+  static ClipLocalizations of(BuildContext context) {
+    return Localizations.of<ClipLocalizations>(context, ClipLocalizations);
+  }
+
+  static const LocalizationsDelegate<ClipLocalizations> delegate = _ClipLocalizationsDelegate();
+
+  /// A list of this localizations delegate along with the default localizations
+  /// delegates.
+  ///
+  /// Returns a list of localizations delegates containing this delegate along with
+  /// GlobalMaterialLocalizations.delegate, GlobalCupertinoLocalizations.delegate,
+  /// and GlobalWidgetsLocalizations.delegate.
+  ///
+  /// Additional delegates can be added by appending to this list in
+  /// MaterialApp. This list does not have to be used at all if a custom list
+  /// of delegates is preferred or required.
+  static const List<LocalizationsDelegate<dynamic>> localizationsDelegates = <LocalizationsDelegate<dynamic>>[
+    delegate,
+    GlobalMaterialLocalizations.delegate,
+    GlobalCupertinoLocalizations.delegate,
+    GlobalWidgetsLocalizations.delegate,
+  ];
+
+  /// A list of this localizations delegate's supported locales.
+  static const List<Locale> supportedLocales = <Locale>[
+    Locale('ar'),
+    Locale('en')
+  ];
+
+  /// No description provided for @gallery.
+  ///
+  /// In en, this message translates to:
+  /// **'Gallery'**
+  String get gallery;
+
+  /// No description provided for @camera.
+  ///
+  /// In en, this message translates to:
+  /// **'Camera'**
+  String get camera;
+
+  /// No description provided for @zoom.
+  ///
+  /// In en, this message translates to:
+  /// **'Zoom'**
+  String get zoom;
+
+  /// No description provided for @remove.
+  ///
+  /// In en, this message translates to:
+  /// **'Remove'**
+  String get remove;
+}
+
+class _ClipLocalizationsDelegate extends LocalizationsDelegate<ClipLocalizations> {
+  const _ClipLocalizationsDelegate();
+
+  @override
+  Future<ClipLocalizations> load(Locale locale) {
+    return SynchronousFuture<ClipLocalizations>(_lookupClipLocalizations(locale));
+  }
+
+  @override
+  bool isSupported(Locale locale) => <String>['ar', 'en'].contains(locale.languageCode);
+
+  @override
+  bool shouldReload(_ClipLocalizationsDelegate old) => false;
+}
+
+ClipLocalizations _lookupClipLocalizations(Locale locale) {
+  
+
+
+// Lookup logic when only language code is specified.
+switch (locale.languageCode) {
+  case 'ar': return ClipLocalizationsAr();
+    case 'en': return ClipLocalizationsEn();
+}
+
+
+  throw FlutterError(
+    'ClipLocalizations.delegate failed to load unsupported locale "$locale". This is likely '
+    'an issue with the localizations generation tool. Please file an issue '
+    'on GitHub with a reproducible sample app and the gen-l10n configuration '
+    'that was used.'
+  );
+}