+++ lib/l10n/clip_localizations_ar.dart
@@ -1,0 +1,23 @@
+
+// ignore: unused_import
+import 'package:intl/intl.dart' as intl;
+import 'clip_localizations.dart';
+
+// ignore_for_file: unnecessary_brace_in_string_interps
+
+/// The translations for Arabic (`ar`).
+class ClipLocalizationsAr extends ClipLocalizations {
+  ClipLocalizationsAr([String locale = 'ar']) : super(locale);
+
+  @override
+  String get gallery => 'المعرض';
+
+  @override
+  String get camera => 'الكاميرا';
+
+  @override
+  String get zoom => 'تكبير';
+
+  @override
+  String get remove => 'ازالة';
+}