+++ lib/l10n/clip_localizations_en.dart
@@ -1,0 +1,23 @@
+
+// ignore: unused_import
+import 'package:intl/intl.dart' as intl;
+import 'clip_localizations.dart';
+
+// ignore_for_file: unnecessary_brace_in_string_interps
+
+/// The translations for English (`en`).
+class ClipLocalizationsEn extends ClipLocalizations {
+  ClipLocalizationsEn([String locale = 'en']) : super(locale);
+
+  @override
+  String get gallery => 'Gallery';
+
+  @override
+  String get camera => 'Camera';
+
+  @override
+  String get zoom => 'Zoom';
+
+  @override
+  String get remove => 'Remove';
+}