+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.0.2
+
+- Use default flutter localizations.
+- Return pickedImage object instead of Uint8List.
+
 # 1.0.1
 
 - Migrate dependencies to nullsaftey.